+++ packages/mwp-tracking-plugin/src/util/trackingUtils.js
